1. New namespace `Set` craeted. 2. Method `Single::distinct()` moved into namespace `Set`. 3. `DistinctTest` moved into `Set` directory. 4. Tests of `Stream::distinct()` moved into separate test `Stream\SetTest`. 5. README updated.
