+++ src/Set.php
@@ -1,0 +1,43 @@
+<?php
+
+declare(strict_types=1);
+
+namespace IterTools;
+
+use IterTools\Util\UniqueExtractor;
+
+class Set
+{
+    /**
+     * Filter out elements from the iterable only returning unique elements.
+     *
+     * If $strict is true:
+     *  - scalars: compares strictly by type;
+     *  - objects: always treats different instances as not equal to each other;
+     *  - arrays: compares serialized.
+     *
+     * If $strict is false:
+     *  - scalars: compares non-strictly by value;
+     *  - objects: compares serialized;
+     *  - arrays: compares serialized.
+     *
+     * @template T
+     * @param iterable<T> $data
+     * @param bool $strict
+     *
+     * @return \Generator<T>
+     */
+    public static function distinct(iterable $data, bool $strict = true): \Generator
+    {
+        $map = [];
+
+        foreach ($data as $datum) {
+            $hash = UniqueExtractor::getString($datum, $strict);
+
+            if (!isset($map[$hash])) {
+                $map[$hash] = true;
+                yield $datum;
+            }
+        }
+    }
+}
