+++ tests/Stream/SingleTest.php
@@ -405,27 +405,6 @@ class SingleTest extends \PHPUnit\Framework\TestCase
                     ->toArray(),
                 [2, 3, 4],
             ],
-            [
-                [1, 2, 3, '1', '2', '3'],
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct(false)
-                    ->toArray(),
-                [1, 2, 3],
-            ],
-            [
-                [1, 2, 3, '1', '2', '3'],
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct()
-                    ->toArray(),
-                [1, 2, 3, '1', '2', '3'],
-            ],
-            [
-                [1, 2, 3, '1', '2', '3', 1, '1'],
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct()
-                    ->toArray(),
-                [1, 2, 3, '1', '2', '3'],
-            ],
         ];
     }
 
@@ -788,27 +767,6 @@ class SingleTest extends \PHPUnit\Framework\TestCase
                     ->limit(10)
                     ->toArray(),
                 [1, 2, 3, 1, 2, 3, 1, 2, 3, 1],
-            ],
-            [
-                $gen([1, 2, 3, '1', '2', '3']),
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct(false)
-                    ->toArray(),
-                [1, 2, 3],
-            ],
-            [
-                $gen([1, 2, 3, '1', '2', '3']),
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct()
-                    ->toArray(),
-                [1, 2, 3, '1', '2', '3'],
-            ],
-            [
-                $gen([1, 2, 3, '1', '2', '3', 1, '1']),
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct()
-                    ->toArray(),
-                [1, 2, 3, '1', '2', '3'],
             ],
         ];
     }
@@ -1173,27 +1131,6 @@ class SingleTest extends \PHPUnit\Framework\TestCase
                     ->toArray(),
                 [1, 2, 3, 1, 2, 3, 1, 2, 3, 1],
             ],
-            [
-                $iter([1, 2, 3, '1', '2', '3']),
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct(false)
-                    ->toArray(),
-                [1, 2, 3],
-            ],
-            [
-                $iter([1, 2, 3, '1', '2', '3']),
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct()
-                    ->toArray(),
-                [1, 2, 3, '1', '2', '3'],
-            ],
-            [
-                $iter([1, 2, 3, '1', '2', '3', 1, '1']),
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct()
-                    ->toArray(),
-                [1, 2, 3, '1', '2', '3'],
-            ],
         ];
     }
 
@@ -1556,27 +1493,6 @@ class SingleTest extends \PHPUnit\Framework\TestCase
                     ->limit(10)
                     ->toArray(),
                 [1, 2, 3, 1, 2, 3, 1, 2, 3, 1],
-            ],
-            [
-                $trav([1, 2, 3, '1', '2', '3']),
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct(false)
-                    ->toArray(),
-                [1, 2, 3],
-            ],
-            [
-                $trav([1, 2, 3, '1', '2', '3']),
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct()
-                    ->toArray(),
-                [1, 2, 3, '1', '2', '3'],
-            ],
-            [
-                $trav([1, 2, 3, '1', '2', '3', 1, '1']),
-                fn (iterable $iterable) => Stream::of($iterable)
-                    ->distinct()
-                    ->toArray(),
-                [1, 2, 3, '1', '2', '3'],
             ],
         ];
     }
