+++ tests/phpunit.xml
@@ -10,6 +10,7 @@
       <directory suffix="Test.php">Reduce</directory>
       <directory suffix="Test.php">Single</directory>
       <directory suffix="Test.php">Summary</directory>
+      <directory suffix="Test.php">Set</directory>
       <directory suffix="Test.php">Stream</directory>
       <directory suffix="Test.php">Util</directory>
     </testsuite>
